+++ client/src/_admin/courses/courses_table/courses_table.js
@@ -1,29 +1,33 @@
 import React, { Component } from "react";
 import PropTypes from 'prop-types';
-import { getCourses,startLoading,deleteCourse } from '../../../__actions/courses_actions';
+import { getCourses,deleteCourse } from '../../../__actions/courses_actions';
 import {connect} from 'react-redux';
 import CourseRow from '../course_row/course_row';
+import {Link} from 'react-router-dom';
 
 
 class courses extends Component {
   constructor() {
     super();
-    this.state = {
-      courses: [],
-      loading: true
-    };
     this.deleteCourse = this.deleteCourse.bind(this);
   }
 
   componentDidMount() {
-   this.props.startLoading();
    this.props.getCourses();
   }
 
   deleteCourse(id){
+    if(!this.props.isAdmin){
+      const errMsg = document.querySelector('.admin-table-error-message');
+      errMsg.style.display = "inline-block";
+      setTimeout(()=>{
+        errMsg.style.display = "none";
+      },3000);
+      window.location.assign("#admin-main-block");
+      return;
+    }
     if(!window.confirm('Are you sure you want to delete this course?')) return;
     this.props.deleteCourse(id, () => this.props.getCourses());
-    
   }
 
   render() {
@@ -32,14 +36,17 @@ class courses extends Component {
       <CourseRow 
       key={course.id} 
       course = {course} 
-      loading={this.props.loading} 
       deleteCourse = {this.deleteCourse}/>
     );
 
     return (
-      <div className="admin-table">
-        <div className="admin-courses-header">
+      <div className="admin-table admin-courses-table">
+        <div className="admin-courses-table-header">
           <h2 className="admin-form-header admin-form-header-fix">Courses Table</h2>
+          <Link to="/admin/courses/add" className="admin-courses-table-icon">
+             <i class="fas fa-plus-circle "></i>
+          </Link>
+          <div className="admin-table-error-message">You are not allowed to perform this action</div>
         </div>
         
         <table>
@@ -48,9 +55,9 @@ class courses extends Component {
               <th className="admin-table-medium-col">Picture</th>
               <th className="admin-table-big-col">Name</th>
               <th className="admin-table-big-col">Instructor</th>
-              <th className="admin-table-medium-col">Field</th>
-              <th className="admin-table-small-col">Points</th>
-              <th className="admin-table-small-col">Id</th>
+              <th className="admin-table-medium-col hide-on-phone">Field</th>
+              <th className="admin-table-small-col hide-on-phone">Points</th>
+              <th className="admin-table-small-col hide-on-phone">Id</th>
               <th className="admin-table-small-col">Delete</th>
             </tr>
           </thead>
@@ -71,8 +78,9 @@ courses.propTypes = {
 
 function mapStateTopProps(state){
   return {
-    courses: state.courses.coursesList
+    courses: state.courses.coursesList,
+    isAdmin: state.auth.isAdmin
   };
 }
 
-export default connect(mapStateTopProps,{getCourses,deleteCourse,startLoading})(courses);
+export default connect(mapStateTopProps,{getCourses,deleteCourse})(courses);
