+++ client/src/_admin/users/users_table/users_table.js
@@ -1,27 +1,32 @@
 import React, { Component } from 'react';
 import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
-import {getUsers, startLoading, deleteUser} from '../../../__actions/users_actions';
-
+import {getUsers, deleteUser} from '../../../__actions/users_actions';
+import {Link} from 'react-router-dom';
 import UserRow from '../user_row/user_row';
 
 class users extends Component {
 
   constructor(){
     super();
-    this.state = {
-      users : [],
-      loading : true
-    };
     this.deleteUser = this.deleteUser.bind(this);
   }
 
   componentDidMount(){
-    this.props.startLoading();
     this.props.getUsers();
   }
 
   deleteUser(id){
+
+    if(!this.props.isAdmin){
+      const errMsg = document.querySelector('.admin-table-error-message');
+      errMsg.style.display = "inline-block";
+      setTimeout(()=>{
+        errMsg.style.display = "none";
+      },3000);
+      window.location.assign("#admin-main-block");
+      return;
+    }
     if(!window.confirm('Are you sure you want to delete this user?')) return;
     this.props.deleteUser(id, ()=> this.props.getUsers());
     
@@ -33,24 +38,27 @@ class users extends Component {
       <UserRow 
       key={user.id} 
       user ={user} 
-      loading={this.props.loading} 
       deleteUser = {this.deleteUser}
       />
     );
 
     return (
-      <div className="admin-table">
+      <div className="admin-table admin-users-table">
         <div className="admin-users-table-header">
           <h2 className="admin-form-header admin-form-header-fix">Users Table</h2>
+          <Link to="/admin/users/add" className="admin-users-table-icon">
+             <i class="fas fa-plus-circle "></i>
+          </Link>
+          <div className="admin-table-error-message">You are not allowed to perform this action</div>
         </div>
         <table>
           <thead>
             <tr>
               <th className="admin-table-medium-col">Image</th>
               <th className="admin-table-big-col">Name</th>
-              <th className="admin-table-big-col">Email</th>
+              <th className="admin-table-big-col hide-on-phone">Email</th>
               <th className="admin-table-medium-col">Role</th>
-              <th className="admin-table-small-col">Id</th>
+              <th className="admin-table-small-col hide-on-phone">Id</th>
               <th className="admin-table-small-col">Delete</th>
             </tr>
           </thead>
@@ -71,8 +79,9 @@ users.propTypes = {
 
 function mapStateTopProps(state){
   return {
-    users: state.users.users
+    users: state.users.users,
+    isAdmin: state.auth.isAdmin
   };
 }
 
-export default connect(mapStateTopProps,{getUsers,startLoading,deleteUser})(users);
+export default connect(mapStateTopProps,{getUsers,deleteUser})(users);
